+++ src/cog/vm/virtual_machine.h
@@ -14,7 +14,7 @@ class virtual_machine {
 private:
     std::stack<value> stack;
     bool allow_run;
-    unsigned int program_counter;
+    size_t program_counter;
 
 public:
     void execute(std::vector<value>& heap, const code_buffer& code, size_t pc, const verbs::verb_table& verbTable, void* system_ptr);
@@ -33,7 +33,7 @@ public:
         return val;
     }
 
-    inline unsigned int get_program_counter() const {
+    inline size_t get_program_counter() const {
         return program_counter;
     }
 };
