+++ src/content/assets/level_adjoin.h
@@ -9,7 +9,7 @@ namespace assets {
 class level_adjoin {
 public:
     flag_set<flags::adjoin_flag> flags;
-    size_t mirror;
+    int mirror;
     float distance;
 };
