+++ src/game/world/camera/camera_model.h
@@ -34,7 +34,7 @@ public:
 
     vector<3> waggle = make_zero_vector<3, float>();
     float waggle_speed = 0.0f;
-    double waggle_time = 0.0;
+    float waggle_time = 0.0;
 };
 
 }
